Media fix project path access using Paths.get('') instead user.dir property

+++ Media/src/main/java/com/cesar/Media/config/MediaCleanUpConfiguration.java
@@ -13,17 +13,10 @@ import jakarta.annotation.PreDestroy;
 @Component
 public class MediaCleanUpConfiguration {
 	
-	
-	private final String MEDIA_DIR_NAME;
-
-	public MediaCleanUpConfiguration(@Value("${media.dirName}") String MEDIA_DIR_NAME) {
-		this.MEDIA_DIR_NAME = MEDIA_DIR_NAME;
-	}
-	
 	@PreDestroy
 	public void cleanUpMediaDirectory() {
 		
-		Path path = Paths.get(MEDIA_DIR_NAME);
+		Path path = Paths.get("").toAbsolutePath().resolve("Media/" + MEDIA_DIR_NAME);
 		
 		if(Files.exists(path)) {
 
@@ -33,11 +26,17 @@ public class MediaCleanUpConfiguration {
 			
 			if(files != null) {
 				for(File file : files) {
-					file.delete();
+					if(file.getPath().endsWith(".png") || file.getPath().endsWith(".jpg")){
+						file.delete();
+					}
 				}
 			}
 		}
-			
 	}
-	
+
+	public MediaCleanUpConfiguration(@Value("${media.dirName}") String MEDIA_DIR_NAME) {
+		this.MEDIA_DIR_NAME = MEDIA_DIR_NAME;
+	}
+
+	private final String MEDIA_DIR_NAME;
 }
